CustomCraft2BasePieces - Initial cleanup
